cleaned up initialize
